Move ConfigShortcuts functions to ShortcutsWidget
